Working convolved heatmap fxn

+++ function_folder/video_making.py
@@ -22,14 +22,12 @@ def export_labelled_frames(df, vname, frame_val, output_dir, investigation = Tru
 	height = int(video.get(cv2.CAP_PROP_FRAME_HEIGHT))
 	fps = int(video.get(cv2.CAP_PROP_FPS))
 
-	print(str(vname).split('/')[-1])
 	
 	## while loop through every frame of the video and label each frame
 	success, image = video.read()
 	count = 0
 	pbar = tqdm(total=frames)
 	while success:
-		pbar.update(count)
 		nose_coords = (int(df['nose']['x'].loc[count]), int(df['nose']['y'].loc[count]))
 		midpoint_coords = (int((df['right ear']['x'].loc[count] + df['left ear']['x'].loc[count]) / 2) , int((df['right ear']['y'].loc[count] + df['left ear']['y'].loc[count]) / 2))
 		if frame_val[count] == 'Somewhere else':
@@ -43,14 +41,15 @@ def export_labelled_frames(df, vname, frame_val, output_dir, investigation = Tru
 		if frame_val[count] == 'Y Investigation':
 			color = (0, 255, 0)
 		image_new = cv2.line(image, nose_coords, midpoint_coords, color, 4)
-		cv2.imwrite(filename = os.path.join(output_dir, str(vname).split('/')[-1] + '_frame_' + str(count) + '.png'), img = image_new)
+		cv2.imwrite(filename = os.path.join(output_dir,'frame_' + str(count) + '.png'), img = image_new)
 		success,image = video.read()
 		count += 1
+		pbar.update(1)
 	pbar.close()
 
 	
 ###also video_making.py
-def ffmpeg_make_video(main_dir, labelled_frames_direc, vname):
+def ffmpeg_make_video(main_dir, labelled_frames_direc, vname, clear_dir = True):
 	import os
 	import subprocess
 	import sys
@@ -71,5 +70,6 @@ def ffmpeg_make_video(main_dir, labelled_frames_direc, vname):
 	### clearing all the images from the hard drive
 	pngs = os.listdir(labelled_frames_direc)
 
-	for img in pngs:
-		os.remove(os.path.join(labelled_frames_direc,img))
+	if clear_dir:
+		for img in pngs:
+			os.remove(os.path.join(labelled_frames_direc,img))
